ADMIN FOR PLACEMENT MANAGEMENT SYSTEM PROJECT

+++ adminmodule/src/main/java/com/cg/admin/AdminRepository.java
@@ -3,7 +3,7 @@ package com.cg.admin;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AdminRepository extends JpaRepository<Admin,Long>
+public interface AdminRepository extends JpaRepository<Admin,Integer>
 {
 
 }
